FastOcc v3: 6가지 병목 해소 + 2단계 커리큘럼 학습
Fix1: dataset_nuscenes_v5.py — numpy 벡터화 + GT .npy 캐싱 (I/O 병목 제거)
Fix2: num_workers=4, gc.collect() 에포크 단위로 제한 (GPU Starvation 해소)
Fix3: model_fastocc.py grid_sample FP32 캐스팅 (FP16 NaN/underflow 방지)
Fix4: gradient accumulation 마지막 배치 처리 누락 수정
Fix5: box_dilate=0 (Lovász Loss FP 페널티 방지)
Fix6: Binary Pre-training(50ep) → 5-class Fine-tuning(150ep) 커리큘럼
- .gitignore 수정: best_fastocc*.pth 예외 순서 fix

+++ code3/model_fastocc.py
@@ -191,11 +191,14 @@ class MultiCamVoxelSampler(nn.Module):
         grid = torch.stack([u_n, v_n], dim=-1)            # (B, N, M, 2)
         grid = grid.reshape(B * N, 1, -1, 2)              # (B*N, 1, M, 2)
 
+        # ★ Fix: FP16 autocast 환경에서 grid_sample 기울기 언더플로우 방지
+        # PyTorch FP16으로 좌표가 작아지면 gradient→0/NaN 발생 (Ped 등 희소 클래스 mIoU=0 원인)
+        # feats와 grid 모두 float32로 명시 캐스팅 후 샘플링
         sampled = F.grid_sample(
-            feats_flat, grid,
+            feats_flat.float(), grid.float(),
             mode='bilinear',
             padding_mode='zeros',
-            align_corners=True)                            # (B*N, C, 1, M)
+            align_corners=True)                            # (B*N, C, 1, M) float32
         sampled = sampled.reshape(B, N, C, -1)             # (B, N, C, M)
 
         # 유효하지 않은 위치 마스킹
